pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DockSuiteX"
version = "1.0.0"
description = "All-in-one protein-ligand docking package for Windows with bundled tools"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
  {name = "Mangalam Gautam Sinha", email = "mangalamgsinha@gmail.com"}
]
keywords = ["docking", "autodock", "vina", "bioinformatics", "ligand", "protein", "drug discovery"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Bio-Informatics"
]
dependencies = [
  "pdbfixer-wheel>=1.11.0",
  "OpenMM>=8.3.1",
  "pandas>=1.3",
  "py3Dmol>=1.8",
  "requests>=2.26"
]

[project.urls]
"Homepage" = "https://github.com/MangalamGSinha/DockSuiteX"
"Documentation" = "https://github.com/MangalamGSinha/DockSuiteX/wiki"
"Source" = "https://github.com/MangalamGSinha/DockSuiteX"
"Issue Tracker" = "https://github.com/MangalamGSinha/DockSuiteX/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

[tool.setuptools.package-data]
"docksuitex" = ["bin/**/*"]
"*" = ["examples/**/*"]

# Optional, only if you have a CLI
# [project.scripts]
# DockSuiteX = "DockSuiteX.cli:main"
